% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.curlStatus}
\alias{.curlStatus}
\title{Determine whether libcurl and libcrypto were found during package 
installation}
\usage{
.curlStatus()
}
\description{
Determine whether libcurl and libcrypto were found during package 
installation
}
\keyword{internal}
