% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.getSzipLoc}
\alias{.getSzipLoc}
\title{If we compiled our own version of SZIP this returns the link path
Othwise it returns an empty string.}
\usage{
.getSzipLoc(path)
}
\description{
If we compiled our own version of SZIP this returns the link path
Othwise it returns an empty string.
}
\keyword{internal}
