% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.getDynamicLinks}
\alias{.getDynamicLinks}
\title{Return the link flags depending upon the status of curl
determined during package installation}
\usage{
.getDynamicLinks()
}
\description{
Return the link flags depending upon the status of curl
determined during package installation
}
\keyword{internal}
