% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.getDynamicLinks}
\alias{.getDynamicLinks}
\title{Return the link flags determined when HDF5 was configured}
\usage{
.getDynamicLinks()
}
\description{
Return the link flags determined when HDF5 was configured
}
\keyword{internal}
