% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{getHdf5Version}
\alias{getHdf5Version}
\title{Report the version of HDF5 distributed with this package}
\usage{
getHdf5Version()
}
\value{
Returns a \code{character} vector of length 1 containing the version
number.
}
\description{
This function returns the version number of the HDF5 library that is 
distributed with this package.
}
\examples{
getHdf5Version()
}
