% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhdf5lib.R
\docType{package}
\name{Rhdf5lib}
\alias{Rhdf5lib-package}
\alias{Rhdf5lib}
\title{Rhdf5lib: A version of the HDF5 library built into an R package.}
\description{
This package provides a compiled version of the HDF5 library bundled within
the R package structure.
It is primarily useful to developers of other R packages who want
to make use of the capabilities of the HDF5 library directly in the C or 
C++ code of their own packages, rather than using a higher level interface.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Huber-group-EMBL/Rhdf5lib}
  \item Report bugs at \url{https://github.com/Huber-group-EMBL/Rhdf5lib/issues}
}

}
\author{
\strong{Maintainer}: Mike Smith \email{grimbough@gmail.com} (\href{https://orcid.org/0000-0002-7800-3848}{ORCID}) [contributor]

Other contributors:
\itemize{
  \item The HDF Group [copyright holder]
}

}
