% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhdf5lib.R
\docType{package}
\name{Rhdf5lib}
\alias{Rhdf5lib}
\title{Rhdf5lib: A version of the HDF5 library built into an R package.}
\description{
This package provides a compiled version of the HDF5 library bundled within
the R package structure.
It is primarily useful to developers of other R packages who want
to make use of the capabilities of the HDF5 library directly in the C or 
C++ code of their own packages, rather than using a higher level interface.
}
